pyproject.toml
[project]
name = "ecommerce-product-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4==4.13.5",
    "ddgs==9.6.0",
    "fastapi==0.116.1",
    "html5lib==1.1",
    "ipykernel>=7.1.0",
    "jinja2==3.1.6",
    "langchain>=1.0.5",
    "langchain-astradb>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=3.0.2",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "lxml==6.0.1",
    "mcp>=1.21.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "ragas==0.3.4",
    "selenium==4.35.0",
    "streamlit==1.49.1",
    "structlog==25.4.0",
    "undetected-chromedriver==3.5.5",
    "uvicorn==0.35.0",
]


authors = [{ name = "Abhimanyu" }]
license = { text = "Proprietary" }

# Optional: CLI command (python -m pip install -e . ke baad yeh create hoga)
[project.scripts]
ecomm-assistant = "ecommerce_prod_assistant.cli:main"

[tool.setuptools.packages.find]
where = ["prod_assistant"]
include = ["ecommerce_prod_assistant*"]
